requirements.txt
django==3.0.3
django-tempus-dominus==5.1.2.17
django-widget-tweaks==1.4.12
